Insert, rather than push, new adapters
The old way of pushing adapters meant that you would have to forcibly
request any kind of provider other than the localStorage one with a
window-based fallback; you could not add other providers in order of
preference, for example, because if the code failed on the adapter
provided in options, it would just plain fail.  This change allows the
user to set up a preferred-order list, ending with localStorage and
window.