Fix remove() bugs in memory and window-name adapters.
If the object passed in to remove() was either not present, or else was
a complete object (from which we were supposed to fetch the 'key' field),
the memory and window-name adapters were erroneously removing the last
entry from the index array, rather than nothing or the correct entry
(respectively).  Add an additional unit test to show this problem.