dom adapter: fix get(array) with nonexistent key; fix remove(array).
Passing an array to get() could previously return a result array shorter
than the parameter array, since keys not present in the lawnchair were missing
from the result.  Add 'null' in the place of the missing result, to be
consistent with other adapters.

Implement batch remove (using code borrowed from the indexedDB adapter).