webkit-sqlite adapter: various fixes to batch(), get(), and remove().
Update the batch() implementation to match that used in the other
adapters.  Removes the need for setInterval() to poll for completion.

Fix sql quoting issues in get() implementation, and ensure that the
order of returned results matches the order of the keys provided as
arguments in the batch get case.

Implement batch remove.  Update the "is this a key or an object containing
a key" test to match that used in the other adapters.